// frontend/app/data/models/user.ts
export class User {
    _id: string;
    name: string;
    email: string;
    username: string;
    password: string;
    confirmPassword: string;
    itemCreatedTimestamp: Date;
    itemCreatedByUser: any;
    itemUpdatedTimestamp: Date;
    itemUpdatedByUser: any;
    itemDeactivedTimestamp: Date;
    itemDeactivedByUser: any;
    itemId: string;
}